Refactor default holder implementation
Extract interface descripbing ability to bind item to view
